+++ test/unit/package_admin/code_coverage.sh
@@ -22,6 +22,7 @@ coverage run -a --source=package_admin test/unit/package_admin/list_repo.py
 coverage run -a --source=package_admin test/unit/package_admin/mail_data.py
 coverage run -a --source=package_admin test/unit/package_admin/main.py
 coverage run -a --source=package_admin test/unit/package_admin/mongo_insert.py
+coverage run -a --source=package_admin test/unit/package_admin/output_run.py
 coverage run -a --source=package_admin test/unit/package_admin/process_yum.py
 coverage run -a --source=package_admin test/unit/package_admin/rabbitmq_publish.py
 coverage run -a --source=package_admin test/unit/package_admin/run_program.py