+++ test/unit/package_admin/kernel_run3.py
@@ -38,7 +38,6 @@ class ArgParser(object):
 
     Methods:
         __init__
-        get_val
 
     """
 
@@ -54,18 +53,6 @@ class ArgParser(object):
 
         self.cmdline = None
         self.args_array = {"-i": "Database_Name:Table_Name"}
-
-    def get_val(self, skey, def_val=None):
-
-        """Method:  get_val
-
-        Description:  Method stub holder for gen_class.ArgParser.get_val.
-
-        Arguments:
-
-        """
-
-        return self.args_array.get(skey, def_val)
 
 
 class Dnf(object):
@@ -100,11 +87,6 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_mongo_rabbit_failure
-        test_rabbit_failure
-        test_rabbit_successful
-        test_mongo_failure
-        test_mongo_successful
         test_kernel_failure
         test_kernel_successful
         test_python_30
@@ -128,130 +110,9 @@ class UnitTest(unittest.TestCase):
 
         self.status = (False, "Kernel_Message")
         self.status2 = (True, None)
-        self.status3 = (False, "Mongo_Message")
-        self.status4 = (False, "Rabbit_Message")
 
         self.results = (False, "Kernel_Message")
         self.results2 = (True, None)
-        self.results3 = (False, "Mongo_Message")
-        self.results4 = (False, "Rabbit_Message")
-        self.results5 = (
-            False, "MongoDB: Mongo_Message RabbitMQ: Rabbit_Message")
-
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
-    @mock.patch("package_admin.kernel_check")
-    def test_mongo_rabbit_failure(self, mock_chk, mock_mongo, mock_rabbit):
-
-        """Function:  test_mongo_rabbit_failure
-
-        Description:  Test with mongo insert and rabbitmq publication failure.
-
-        Arguments:
-
-        """
-
-        mock_rabbit.return_value = self.status4
-        mock_mongo.return_value = self.status3
-        mock_chk.return_value = self.status2, self.data
-
-        self.assertEqual(
-            package_admin.kernel_run(self.args, self.dnf), self.results5)
-
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
-    @mock.patch("package_admin.kernel_check")
-    def test_rabbit_failure(self, mock_chk, mock_mongo, mock_rabbit):
-
-        """Function:  test_rabbit_failure
-
-        Description:  Test with rabbitmq publication failure.
-
-        Arguments:
-
-        """
-
-        mock_rabbit.return_value = self.status4
-        mock_mongo.return_value = self.status2
-        mock_chk.return_value = self.status2, self.data
-
-        self.assertEqual(
-            package_admin.kernel_run(self.args, self.dnf), self.results4)
-
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
-    @mock.patch("package_admin.kernel_check")
-    def test_rabbit_successful(self, mock_chk, mock_mongo, mock_rabbit):
-
-        """Function:  test_rabbit_successful
-
-        Description:  Test with rabbitmq publication successful.
-
-        Arguments:
-
-        """
-
-        mock_rabbit.return_value = self.status2
-        mock_mongo.return_value = self.status2
-        mock_chk.return_value = self.status2, self.data
-
-        self.assertEqual(
-            package_admin.kernel_run(self.args, self.dnf), self.results2)
-
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
-    @mock.patch("package_admin.kernel_check")
-    def test_mongo_failure(self, mock_chk, mock_mongo, mock_rabbit):
-
-        """Function:  test_mongo_failure
-
-        Description:  Test with mongo insert failure.
-
-        Arguments:
-
-        """
-
-        mock_rabbit.return_value = self.status2
-        mock_mongo.return_value = self.status3
-        mock_chk.return_value = self.status2, self.data
-
-        self.assertEqual(
-            package_admin.kernel_run(self.args, self.dnf), self.results3)
-
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
-    @mock.patch("package_admin.kernel_check")
-    def test_mongo_successful(self, mock_chk, mock_mongo, mock_rabbit):
-
-        """Function:  test_mongo_successful
-
-        Description:  Test with mongo insert successful.
-
-        Arguments:
-
-        """
-
-        mock_rabbit.return_value = self.status2
-        mock_mongo.return_value = self.status2
-        mock_chk.return_value = self.status2, self.data
-
-        self.assertEqual(
-            package_admin.kernel_run(self.args, self.dnf), self.results2)
 
     @mock.patch("package_admin.kernel_check")
     def test_kernel_failure(self, mock_chk):
@@ -269,13 +130,9 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             package_admin.kernel_run(self.args, self.dnf), self.results)
 
-    @mock.patch("package_admin.mail_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.display_data", mock.Mock(return_value=True))
-    @mock.patch("package_admin.write_file", mock.Mock(return_value=True))
-    @mock.patch("package_admin.rabbitmq_publish")
-    @mock.patch("package_admin.mongo_insert")
+    @mock.patch("package_admin.output_run")
     @mock.patch("package_admin.kernel_check")
-    def test_kernel_successful(self, mock_chk, mock_mongo, mock_rabbit):
+    def test_kernel_successful(self, mock_chk, mock_output):
 
         """Function:  test_kernel_successful
 
@@ -285,8 +142,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_rabbit.return_value = self.status2
-        mock_mongo.return_value = self.status2
+        mock_output.return_value = self.status2
         mock_chk.return_value = self.status2, self.data
 
         self.assertEqual(